Batch Grading: Added support for grading multiple submissions in a single run
