+++ app/scripts/cli/mvp_web.py
@@ -23,7 +23,7 @@ DEFAULT_LECTURE_CHUNKS = (
     / "describe_openai_gpt-4o-2024-11-20_v2_semantic"
     / "chunks.jsonl"
 )
-DEFAULT_RUBRIC_ROOT = PROJECT_ROOT / "rubrics"
+DEFAULT_RUBRIC_ROOT = Path("/Users/sai/Downloads/Spring 2026 2/Assignment Rubrics")
 
 MODEL_BY_PROVIDER = {
     "openai": "gpt-4o-2024-11-20",
