--- app/scripts/requirements.txt
@@ -1,19 +1,0 @@
-pandas>=2.0.0
-openpyxl>=3.1.0
-pypdf>=3.17,<4.0
-Pillow>=10.0.0
-matplotlib>=3.8.0
-pytesseract>=0.3.10
-opencv-python-headless>=4.10.0
-pypdfium2>=4.30.0
-PyMuPDF>=1.24.0
-anthropic>=0.39.0
-openai>=1.35.0
-chromadb>=0.5.0
-python-docx>=1.1.2
-beautifulsoup4>=4.12.0
-python-pptx>=1.0.2
-
-camelot-py>=0.11.0
-python-dotenv>=1.0.0
-flask>=3.0.0
