+++ app/scripts/storage/jsonl_writer.py
@@ -1,30 +1,4 @@
-from __future__ import annotations
+# Re-exported from core.io — do not add logic here.
+from core.io import write_json, write_jsonl, write_per_file_json
 
-import json
-from pathlib import Path
-from typing import Any
-
-
-def write_jsonl(path: Path, rows: list[dict[str, Any]]) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with path.open("w", encoding="utf-8") as f:
-        for row in rows:
-            f.write(json.dumps(row, ensure_ascii=True) + "\n")
-
-
-def write_json(path: Path, obj: dict[str, Any] | list[Any]) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(obj, indent=2, ensure_ascii=True), encoding="utf-8")
-
-
-def write_per_file_json(per_file_root: Path, rel_path: str, payload: dict[str, Any]) -> Path:
-    rel = Path(rel_path)
-    if rel.is_absolute() or any(part == ".." for part in rel.parts):
-        raise ValueError("rel_path must be a safe relative path")
-    out_path = per_file_root / f"{rel}.extraction.json"
-    root = per_file_root.resolve()
-    if not out_path.resolve().is_relative_to(root):
-        raise ValueError("rel_path escapes per_file_root")
-    out_path.parent.mkdir(parents=True, exist_ok=True)
-    out_path.write_text(json.dumps(payload, indent=2, ensure_ascii=True), encoding="utf-8")
-    return out_path
+__all__ = ["write_jsonl", "write_json", "write_per_file_json"]
